requirements.txt
PyPDF2
pytesseract
Pillow
python-docx
